I1-76 ref: rename folder
