+++ src/widgets/ForgotPassword/index.ts
@@ -1,0 +1,1 @@
+export { ForgotPassword } from '@/widgets/ForgotPassword/ui/ForgotPassword'
