+++ src/widgets/ForgotPassword/ui/ForgotPassword.module.scss
